Game 2001 Commit 1
First commit to the repository for Game 2001 a2. This is a bit late than intended as I didn't notice the repository part. So far in the project a base Linked List has been created that has a node class and an iterator class. The linked list is also tested to work in the main. The priority queue is being created now.